+++ pubspec.yaml
@@ -8,7 +8,7 @@ environment:
   sdk: '>=1.20.1 <2.0.0'
 
 dependencies:
-  vm_service_lib: ^0.3.0
+  vm_service_lib: ^0.3.3
 
 dev_dependencies:
   browser: ^0.10.0
@@ -20,7 +20,3 @@ web:
 
 transformers:
 - dart_to_js_script_rewriter
-
-dependency_overrides:
-  vm_service_lib:
-    path: /Users/devoncarew/projects/dart-lang/vm_service_drivers/dart
