+++ web/index.html
@@ -31,7 +31,7 @@
         <a class="masthead-logo" id="title">Observatory</a>
 
         <nav class="masthead-nav" id="main-nav">
-            <a href="/">Main</a>
+            <a href="/">Overview</a>
             <a href="/timeline"><span class="octicon octicon-pulse"></span> Timeline</a>
             <a href="/performance"><span class="octicon octicon-dashboard"></span> Performance</a>
             <a href="/memory"><span class="octicon octicon-package"></span> Memory</a>
