+++ web/main.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:perf_tool/main.dart';
+import 'package:perf_tool/perf_main.dart';
 
 // TODO: we connect to a VM instance, by port
