+++ web/styles.css
@@ -69,6 +69,11 @@ h2 {
     border-bottom: 1px solid #eee;
     margin-bottom: 8px;
     padding-bottom: 0px;
+    margin-top: 20px;
+}
+
+h2:first-of-type {
+    margin-top: 0;
 }
 
 td.right {
